# api-docs/openapi.yaml
openapi: 3.1.0
info:
  title: Kayak Simulation Distributed Services API
  version: 1.0.0
  description: |
    API specification for the Kayak-style distributed system. The platform exposes REST
    endpoints for user management, listings search, bookings, billing, admin operations,
    analytics, and the FastAPI-based agentic recommendation service. All synchronous
    operations occur over HTTP while asynchronous updates are published on Kafka topics
    and relayed to clients via WebSockets.
  contact:
    name: Kayak Simulation Platform Team
    email: eng@kayak-sim.local
servers:
  - url: https://api.kayak-sim.prod
    description: Production API Gateway
  - url: https://api.staging.kayak-sim.local
    description: Staging API Gateway
  - url: http://localhost:8080
    description: Local development gateway
...
paths:
  /api/v1/users:
    get:
      tags: [Users]
      summary: List users
      description: Retrieve users with optional filters. Requires admin privileges.
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: email
          in: query
          schema:
            type: string
            format: email
          description: Filter by email address.
        - name: state
          in: query
          schema:
            $ref: '#/components/schemas/USStateCode'
          description: Filter by state code.
      responses:
        '200':
          description: Paginated user result set.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
    post:
      tags: [Users]
      summary: Create user
      description: Register a new user profile. SSN is optional unless the profile type is a property partner.
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: User created successfully.
          headers:
            Location:
              description: URL of the created user resource.
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Duplicate user conflict (e.g., SSN or email already registered).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/users/{userId}:
    parameters:
      - $ref: '#/components/parameters/UserIdPath'
    get:
      tags: [Users]
      summary: Get user
      description: Retrieve the full profile for a user.
      operationId: getUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User details returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFoundError'
    put:
      tags: [Users]
      summary: Replace user
      description: Update a user's full profile. Supports changing all attributes.
      operationId: replaceUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: Updated user state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      tags: [Users]
      summary: Delete user
      description: Permanently remove a user and anonymize related analytics records.
      operationId: deleteUser
      responses:
        '204':
          description: User deleted successfully.
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /api/v1/users/{userId}/bookings:
    parameters:
      - $ref: '#/components/parameters/UserIdPath'
    get:
      tags: [Users, Bookings]
      summary: List user bookings
      operationId: listUserBookings
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/BookingStatus'
          description: Filter by booking status.
      responses:
        '200':
          description: Paginated user bookings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingListResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
    post:
      tags: [Users, Bookings]
      summary: Create user booking
      description: |
        Create a booking for a user. The operation is transactional: on downstream failure the
        booking is rolled back and a compensating Kafka message is emitted.
      operationId: createUserBooking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingCreateRequest'
      responses:
        '201':
          description: Booking created. Payment remains pending until confirmed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Conflicting booking detected (e.g., double booking of same seat/room).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '424':
          description: Downstream dependency failure; no partial state committed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination (1-indexed).
    PageSizeParam:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 25
      description: Number of items to return per page.
    UserIdPath:
      name: userId
      in: path
      required: true
      description: Unique user identifier (UUID).
      schema:
        type: string
        format: uuid
  responses:
    ValidationError:
      description: Input validation failed; see details array.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
    NotFoundError:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Authentication required.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: Caller lacks necessary permissions.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    PaginationMetadata:
      type: object
      required: [page, pageSize, totalItems, totalPages]
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 200
        totalItems:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code.
        message:
          type: string
          description: Human-readable error description.
        correlationId:
          type: string
          format: uuid
          description: Trace ID for support teams.
        details:
          type: array
          items:
            type: string
          description: Additional error context.
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            errors:
              type: array
              description: Field level validation failures.
              items:
                type: object
                required: [field, message]
                properties:
                  field:
                    type: string
                  message:
                    type: string
                  rejectedValue:
                    description: The value that failed validation.
    Address:
      type: object
      required: [line1, city, state, zipCode]
      properties:
        line1:
          type: string
          maxLength: 255
        line2:
          type: string
          maxLength: 255
        city:
          type: string
          maxLength: 120
        state:
          $ref: '#/components/schemas/USStateCode'
        zipCode:
          type: string
          pattern: '^\\d{5}(-\\d{4})?$'
          description: Supports '#####' or '#####-####'.
    PaymentInstrument:
      type: object
      required: [paymentType, cardLast4]
      properties:
        paymentType:
          type: string
          enum: [credit_card, debit_card, paypal, apple_pay, google_pay]
        cardBrand:
          type: string
          example: VISA
        cardLast4:
          type: string
          pattern: '^\\d{4}$'
        expirationMonth:
          type: integer
          minimum: 1
          maximum: 12
        expirationYear:
          type: integer
          minimum: 2024
        billingAddress:
          $ref: '#/components/schemas/Address'
...
    User:
      type: object
      required:
        - id
        - firstName
        - lastName
        - email
        - phoneNumber
        - address
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        ssn:
          type: string
          pattern: '^\\d{3}-\\d{2}-\\d{4}$'
          nullable: true
          description: Optional SSN stored for property partners. Travelers do not provide this.
        profileType:
          type: string
          enum: [traveler, property_owner]
          default: traveler
        requiresSsn:
          type: boolean
        hasSsnOnFile:
          type: boolean
        compliance:
          $ref: '#/components/schemas/UserCompliance'
        partnerDetails:
          $ref: '#/components/schemas/PartnerProfile'
        firstName:
          type: string
          maxLength: 80
        lastName:
          type: string
          maxLength: 80
        email:
          type: string
          format: email
        phoneNumber:
          type: string
          pattern: '^\\+1-\\d{3}-\\d{3}-\\d{4}$'
          description: Normalized US phone number (+1-XXX-XXX-XXXX).
        address:
          $ref: '#/components/schemas/Address'
        profileImageUrl:
          type: string
          format: uri
        creditCards:
          type: array
          items:
            $ref: '#/components/schemas/PaymentInstrument'
        loyaltyTier:
          type: string
          enum: [none, silver, gold, platinum]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        preferences:
          type: object
          additionalProperties: true
    UserCompliance:
      type: object
      properties:
        requiresSsn:
          type: boolean
          description: Indicates whether the user must provide an SSN before managing properties.
        ssnOnFile:
          type: boolean
          description: True when an SSN has been provided.
        verifiedAt:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when the SSN was last verified.
    PartnerProfile:
      type: object
      description: Business metadata captured for property partners.
      properties:
        companyName:
          type: string
        contactName:
          type: string
        contactEmail:
          type: string
          format: email
        portfolioSize:
          type: integer
          minimum: 0
          nullable: true
        website:
          type: string
          format: uri
    UserCreateRequest:
      type: object
      required:
        - firstName
        - lastName
        - email
        - phoneNumber
        - address
      properties:
        ssn:
          type: string
          pattern: '^\\d{3}-\\d{2}-\\d{4}$'
          nullable: true
          description: Optional during signup. Required later only for property partners.
        profileType:
          type: string
          enum: [traveler, property_owner]
          default: traveler
          description: Travelers can skip SSN. Property partners must supply SSN before listing approvals.
        firstName:
          type: string
          maxLength: 80
        lastName:
          type: string
          maxLength: 80
        email:
          type: string
          format: email
        phoneNumber:
          type: string
          pattern: '^\\+1-\\d{3}-\\d{3}-\\d{4}$'
        address:
          $ref: '#/components/schemas/Address'
        profileImageUrl:
          type: string
          format: uri
        partnerProfile:
          $ref: '#/components/schemas/PartnerProfile'
          description: Required when profileType is `property_owner`.
        paymentInstruments:
          type: array
          items:
            $ref: '#/components/schemas/PaymentInstrument'
        marketingOptIn:
          type: boolean
          default: false
    UserUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/UserCreateRequest'
      required:
        - firstName
        - lastName
        - email
        - phoneNumber
        - address
...
    UserListResponse:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/User'
        pagination:
          $ref: '#/components/schemas/PaginationMetadata'
    FlightClass:
      type: string
      enum: [economy, premium_economy, business, first]
...
    CarType:
      type: string
      enum: [suv, sedan, compact, luxury, convertible, van, truck]
...
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: number
          format: float
          minimum: 0
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          example: USD
    BookingType:
      type: string
      enum: [flight, hotel, car]
    BookingStatus:
      type: string
      enum: [PENDING, CONFIRMED, COMPLETED, CANCELLED, FAILED]
    Booking:
      type: object
      required:
        - id
        - userId
        - bookingType
        - status
        - createdAt
        - updatedAt
        - itinerary
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        bookingType:
          $ref: '#/components/schemas/BookingType'
        status:
          $ref: '#/components/schemas/BookingStatus'
        itinerary:
          $ref: '#/components/schemas/Itinerary'
        price:
          $ref: '#/components/schemas/Money'
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BookingCreateRequest:
      type: object
      required:
        - userId
        - bookingType
        - itinerary
        - quoteExpiresAt
      properties:
        userId:
          type: string
          format: uuid
        bookingType:
          $ref: '#/components/schemas/BookingType'
        itinerary:
          $ref: '#/components/schemas/Itinerary'
        quoteId:
          type: string
        quoteExpiresAt:
          type: string
          format: date-time
        paymentMethodId:
          type: string
          format: uuid
        metadata:
          type: object
          additionalProperties: true
...
    BookingListResponse:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Booking'
        pagination:
          $ref: '#/components/schemas/PaginationMetadata'
    Itinerary:
      type: object
      required: [segments]
      properties:
        segments:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ItinerarySegment'
        guests:
          type: array
          items:
            $ref: '#/components/schemas/Guest'
    ItinerarySegment:
      type: object
      required: [type, referenceId, startDateTime, endDateTime]
      properties:
        type:
          type: string
          enum: [flight, hotel, car]
        referenceId:
          type: string
        startDateTime:
          type: string
          format: date-time
        endDateTime:
          type: string
          format: date-time
        fareClass:
          $ref: '#/components/schemas/FlightClass'
        roomTypeCode:
          type: string
        carType:
          $ref: '#/components/schemas/CarType'
        seatAssignments:
          type: array
          items:
            type: string
        amenityCodes:
          type: array
          items:
            type: string
    Guest:
      type: object
      required: [firstName, lastName]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
          pattern: '^\\+1-\\d{3}-\\d{3}-\\d{4}$'
...
    USStateCode:
      type: string
      enum:
        - AL
        - AK
        - AZ
        - AR
        - CA
        - CO
        - CT
        - DE
        - FL
        - GA
        - HI
        - ID
        - IL
        - IN
        - IA
        - KS
        - KY
        - LA
        - ME
        - MD
        - MA
        - MI
        - MN
        - MS
        - MO
        - MT
        - NE
        - NV
        - NH
        - NJ
        - NM
        - NY
        - NC
        - ND
        - OH
        - OK
        - OR
        - PA
        - RI
        - SC
        - SD
        - TN
        - TX
        - UT
        - VT
        - VA
        - WA
        - WV
        - WI
        - WY
      description: Two-letter postal abbreviation for US states.
...